feat: add support for pages for contracts
